src/authtoken_service/requirements.txt
alembic==1.11.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
build==0.10.0
CacheControl==0.12.14
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.4
crashtest==0.4.1
cx-Oracle==8.3.0
distlib==0.3.7
dulwich==0.21.5
exceptiongroup==1.1.2
filelock==3.12.2
Flask==2.3.2
Flask-API==3.1
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
google==3.0.0
grpcio==1.57.0
grpcio-tools==1.57.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
keyring==23.13.1
load-dotenv==0.1.0
lockfile==0.12.2
Mako==1.2.4
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.5
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pexpect==4.8.0
pipenv==2023.7.23
pkginfo==1.9.6
platformdirs==3.9.1
pluggy==1.2.0
poetry==1.5.1
poetry-core==1.6.1
poetry-plugin-export==1.4.0
protobuf==4.24.0
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
PyMySQL==1.1.0
pyproject_hooks==1.0.0
pytest==7.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rapidfuzz==2.15.1
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.9.2
shellingham==1.5.3
soupsieve==2.4.1
SQLAlchemy==2.0.18
SQLAlchemy-Utils==0.41.1
tomli==2.0.1
tomlkit==0.12.1
trove-classifiers==2023.8.7
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
virtualenv==20.24.2
virtualenv-clone==0.5.7
webencodings==0.5.1
Werkzeug==2.3.6
xattr==0.10.1
zipp==3.16.0
bcrypt==4.0.1
gunicorn==21.2.0
